+++ CHANGELOG.md
@@ -14,6 +14,10 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
 
 - New alias method `sdk.securitydata.savedsearches.search_file_events()` for existing method `sdk.securitydata.savedsearches.execute()`.
 
+- Two new `riskIndicator.UserBehavior` filter values
+    - `First use of destination`
+    - `Rare use of destination`
+
 ### Changed
 
 - Updated minimum version of `requests` library to 2.4.2